backend/requirements.txt
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
boto3>=1.34.0
Flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
openai>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
PyYAML>=6.0.0
# Optional: anthropic>=0.18.0  # Uncomment if using Claude

